Add models for SSH signed commits
This is part one of a multi-part implementation to support validation of
git commits signed by SSH keys. This adds the database tables and models
for SSH commit signatures, using the same pattern as the existing
gpg_signatures and x509_signatures tables. These will be used to show
the verification status on commits which are signed.

Changelog: added