Fix pending members view
The pending members view is part of the groups/usage_quotas page where
we try to initialilze the NamespaceStorage Vue application as well.

However, as we don't render the #js-storage-counter-app element with
the data for the Vue application, `initNamespaceStorage` raises an
error.

To fix this, we don't try to initialize the Vue Application and return
early (as we do it for the other Vue applications as well).