Prepare new Slack events processing API
Link unfurling https://gitlab.com/groups/gitlab-org/-/epics/6389 is the
first Slack feature that will use the Slack events API.

All Slack events that we subscribe to in future will be sent to a single
endpoint of GitLab, regardless of the Slack workspace or event type. So
the endpoint needs to be very performant.

For this first iteration, this commit is to prepare the events endpoint
and process the challenge, but not process any events yet.

https://gitlab.com/gitlab-org/gitlab/-/issues/358676

Changelog: added