Fallback to Diff::File#diff_lines_for_serializer when limit is hit
When parsing conflict file, we have a limit of 200kb. When file
exceeds that size a `Gitlab::Git::Conflict::Parser::UnmergeableFile`
exception gets raised. It's not being handled anywhere so when
we display conflicts on diffs, it errors out the entire diff.

To fix this, we rescue the exception and fallback to the
corresponding diff file's `#diff_lines_for_serializer`. This way,
we'll still be able to show the diff when it occurs.

This is behind `display_merge_conflicts_in_diff` feature flag.