Move pipeline vuln. report store to the entry point
This moves the pipeline vulnerability store's initialization to the
app's entry point. The store is then set up in the app's root component.
This ensures that the app can seamlessly register its store modules and
plugins no matter where/how the store instance has been created. This makes
it easier for us to migrate this Vue app to the unified pipeline tabs
app that will register a single store to which each tab app should
attach its own modules.