Introduce source versions for BulkImport trackers configuration
GitLab Importer aka BulkImport uses a configuration file to define
which pipelines will be executed when importing groups or projects

However, some pipelines can't be executed when importing from sources
instances running a GitLab version incompatible with the pipeline.

This change introduces the attributes miminum_source_version and
maximum_source_version to the pipeline configuration to skip a pipeline
depending on the source instance versions.

When the source instance version doesn't meet the version requirements,
the pipeline trackers are created as skipped, which prevents the
pipeline from being executed.

Changelog: added