Fix group IP restrictions not enforced for container registry requests
Group IP restrictions only apply when the declarative policy checks
are used, but deploy tokens and keys previously did not use these
checks. Previously the existence of an active deploy token or key in a
project's configuration was sufficient to grant access.

We fix this issue by adding support for doing policy checks for a
deploy token and key and using this in the JWT authentication flow.

We don't call `Ability.allowed?` on deploy keys yet, but if we did
we would need to ensure this were set.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/363651

EE: true
Changelog: security