Add and expose Jira app OAuth application ID
This adds a new application setting, so users can enter the ID of
the OAuth application they created to use the GitLab for Jira app
on self-managed.

GitLab.com will serve as a proxy for the app. This means it will need
to fetch the application ID from the self-managed instance in order to
initialize the OAuth login flow. Therefor the application ID setting
is publicly exposed.

Changelog: added