Initial cleanup of testing docs
The unit test report doc was full of examples
and needed to be split into it's own page.
The doc also needed to be in it's own directory
so that the docs can be grouped together logically,
along with the images. This new directory also
needed an index.md, so the testing_and_reports...
doc was moved here and made the index.md. That
doc was not in active use, and not present in
the docs navigation, so it's a good choice.