Fix Members Activate and Await Service
It's possible for membership records to be invalid when the inherited
`access_level` is higher. When using the state machine to transition
between membership states, it would validate the records and therefore
fail to change the state.

To prevent this bug from happening we skip the validation by using
`update_all`. To still guarantee valid data, we moved the validation and
restrictions to the `Members::ActivateService` and
`Members::AwaitService`.

* We now call `UserProjectAccessChangedService` when setting a member to
  be `awaiting`, which runs otherwise as an `after_commit` hook.
* We check the free user cap limits when activating a user.
* We prevent the last owner to be set to awaiting.
* We don't allow a user to set themselves to be `awaiting` anymore.