Run RubySampler and ThreadSampler in Puma primary
We currently have no process and thread metrics for
the puma_master process in Prometheus. This was a
bit tricky to do, since we must not inherit down
sampler state when forking into workers.

By stopping existing samplers and recreating the
instance, this should be safe to do now.

Changelog: changed