Add workhorse internal upload API, observation only
This adds an internal API endpoint for Workhorse upload
preauthorization. This endpoint is only for the catchall Workhorse
route.

The current iteration calls the API and logs the result. We want to
observe how this behaves in production. Prior to this change we had no
visibility into how often this catchall route is used for uploads. If
all goes well we will make the API mandatory in a later commit.

Changelog: other