Always destroy webhooks synchronously
Previously we fell-back to asynchronous batch deletion if there were
many logs, but we no longer need to do that now that we do not have a
foreign key preventing deletion.

Now we can delete the hook in all cases, and then start a background
worker to delete the logs. This is actually unnecessary, since they will
be truncated over time anyway, but it is nice to minimize table bloat.

Changelog: fixed