Preload `avatar_url` for CompareController
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/363879

**Problem**

We load all avatars of forked projects and that causes a performance
problem.

**Solution**

By default we use batch loading for avatars, but in this case we want
to avoid that. I call `avatar_url` method early to change the order of
the batch load and fix the unoptimized query to fetch all avatars of
project's forks.

Changelog: performance