Report warnings from the timebox resolver
If the timebox resolver is not able to produce a result, it adds
an error to the report. If the error is present, none of the
other fields will be.

These errors include a localized message, as well as a static code
indicating the class of error, represented as an enum.

Changelog: changed