Enable escalations of incidents created manually
Incidents created from alerts are automatically escalated to on-call
users via the alert. This change allows
users to escalate incidents which have been created manually by setting
an escalation policy for the incident. The status of the incident can be
used to acknowledge the pages from the escalation.

Changelog: added