Makes exceed_query_limit matcher less noisy
The matcher will currently report query groups, and their differences.
Many groups will differ only in parameters. For example the control
query will have `Project.where(id: 1)` and the test will have
`Project.where(id: [1, 2, 3])`. We want to mark this as a successful
group, and exclude it from the diff, making the diff more useful
when there are failures.