Add GitLab agent image tag to install command
Currently, the commands used to install the GitLab agent will always
install the latest chart version, which, by default, uses the latest
released agent. This could cause problems on older versions of GitLab,
as we only guarantee N-1 version compatibility (new agents will not
always work with old GitLabs). This makes it so that the version we
install is the same as the KAS version (for every KAS version, there is
a corresponding agentk version).

Changelog: fixed