Add terminating newline to /:username.keys and /:username.gpg
This change makes the output compliant with POSIX 3.206 (which defines a
line as a sequence of non-newline characters plus a terminating newline
character) and RFC 4880 section 6.2 (which specifies that the
terminating newline is to be considered part of the ASCII armored key),
respectively, for improved interoperability, e.g. when adding to
authorized_keys files.

Changelog: changed