Log GitHub Commit Status API responses
A recurring support question is around GitHub external
pull request pipeline status not receiving completion updates
despite no errors logged within GitLab.com Kibana.

It becomes difficult to answer with certainty if GitLab
sent the event to GitHub at the completion of the pipeline,
or about the ordering of updates sent to it.

This change adds a log every time we make a Commit Status API
call to GitHub REST endpoints and adds parameters to the log
to help answer the questions above and also make it easier to
find results when searching for a pipeline.

Changelog: added
EE: true