Reduce response cost of deploy keys autocompleter
The autocomplete URL for deploy keys uses a serializer that
attempts to respond with every project the deploy key has been
associated with.

When the list of projects associated to a few eligible keys
exceeds 10-20k projects the autocompleter endpoint
begins to time out and this causes various UIs to break.

For the purpose of the autocompleter only the relevant
key list is required, and the projects are unnecessary.

This change splits the entity into a basic form and an extended
form, with the extended form preserving the current behavior
for other uses of the entity such as APIs.

The basic format will now exclude project list information
which helps the autocompleter respond quicker.

Changelog: performance