Convert SubscriptionHistory create to an allowlist
Previously, the attributes stored to the GitlabSubscriptionHistory table
were selected using a denylist. This caused issues with attributes being
added to this list and then going through the canary stage for
deployment.

Using the denylist approach, we tried to fetch the attributes from the
database directly and then save those to the history table. This meant
that the new attribute (which was deployed to main and canary at the
same time) would attempt to be stored to the history table as the
denylist was only deployed to canary. This caused errors updating
subscriptions in the main stage.

EE: true
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89370
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/360628