Fix `gitlab:db:validate_config` to store `db_config_name` of a database
The previously used `system_identifier` of `pg_control_system()` stays
the same after database promotion and makes the `gitlab:db:validate_config`
to fail since we expect this identifier to be different.

We cannot fix `system_identifer` since it is set on `init xlog`. This
fixes the problem by using an internal AR table to store relevant information.

Since we connect to "write" the writes will be immediate and this way
we can discover if database is being or not shared.

Changelog: fixed