Count awaiting users on checkout free groups
This changes how we calculate the billable members for free plans on the
`free_user_cap` when they purchase a plan.

Example: A free group where `free_user_cap` is enabled has 7 users.
2 users get their memberships set to `awaiting` when `free_user_cap`
applies. When they want to purchase a plan it's likely that they want to
purchase seats for these 2 users. We therefore need to count them in the
checkout process as `billable` members.

We also have to change the way how we calculate `seats_in_use` as it was
using `billable_members_count`. It's now using directly
`billed_user_ids` and we therefore distinguish between billable
(potentially billable) and billed users (actually billed users).