Fix secondary storage namespace for duplicate jobs
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/85740 we
introduced a `MultiStore` to transition duplicate jobs data from
`Sidekiq.redis` to `SharedState`. We used the `Gitlab::Redis::Queues`
store to point to the current data, but omitted that `Sidekiq.redis` is
namespaced. This is actually an existing mishap, in that with the
current state keys are actually stored under
"resque:gitlab:resque:gitlab:duplicate:". We add the namespace to the
Queues store for compatibility with the current state, but leave
`SharedState` with no namepsace so keys that land there have proper
format.