Make LicenseFinder work with platform-specific gems
bundler v2.2+ and LicenseFinder do not play well together when:

1. There are native gems installed (e.g. nokogiri, grpc, and
google-protobuf).

2. The `Gemfile.lock` doesn't list the platform-specific gems that
were installed.

A full explanation is here:
https://github.com/pivotal/LicenseFinder/issues/828#issuecomment-953359134

To work around the issue, we configure bundler to install gems for the
current Ruby platform, which causes Gemfile and Gemfile.lock to be
updated with the platform-specific gems.  This allows LicenseFinder to
run properly. After it finishes, we clean up the mess.