Do not retry queries if a transaction is open
Retrying queries on a new connection when a transaction is open discards
the transaction context without the application knowing. This can lead
to subtle data bugs.

Simply allow the error to propagate in this case, since it's not
possible to recreate the full transaction context on the new database
connection.