Rescue ConflictSideMissing error
`Gitlab::Git::Conflict::Resolver::ConflictSideMissing` error gets
raised when we try to list conflicts but the reference from a
fork isn't present on canonical repo yet (it happens async).

Before this fix, when this error is raised we're not able to
display the diff.

This is part of a feature behind the
`display_merge_conflicts_in_diff` feature flag.