Add re enqueue support in the delete container repository worker
To support the registry migration. Image repositories that are being
migrated can't be deleted. This has unwanted side effects, in particular
on the Container Registry side.

The worker will now check if the image repository is being migrated. If
that's the case, it will compute an estimation for the migration
duration and use that to re_enqueue the job in the future.

This change is behind a feature flag.