CarrierWave file variants are not supported
CarrierWave has a mechanism for having multiple versions of the same
file. For instance a full-size image and a thumbnail. This mechanism
is not compatible with some of GitLab's customizations to CarrierWave.
This commit adds a warning for developers that they should not use the
version mechanism. This does not affect existing GitLab functionality.