Add start_event_timestamp and end_event_timestamp in order_by_duration
This is necessary because
`Gitlab::Analytics::CycleAnalytics::Aggregated::RecordsFetcher`, when
sorting by duration, only selects columns that are part of the ORDER BY
condition. Since we need start_event_timestamp and end_event_timestamp,
we append them to the order condition, even though they do not actually
influence the order.