Allow access to project-level packages API with CI job token
This adds job token authentication to the packages GET, single
package GET, and the package DELETE endpoint, to allow
more flexible workflows with package uploads as outlined in
issue https://gitlab.com/gitlab-org/gitlab/-/issues/349418.
It also adds tests with a shared example to ensure all permisson
level behaviors are as expected, and adds the endpoints that
are now available via the job token in the documentation.

Changelog: added