Add test case for overriding button type
This is a follow-up to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89159.

`Pajamas::ButtonComponent` currently allows the `type` attribute to be
overridden via `button_options`, but it was not covered by a test. This
adds that test.

See for example https://gitlab.com/gitlab-org/gitlab/-/issues/363946
where this is relevant, as there's a `type="submit"` button there.