Fix user cap alert on non-namespace pages
When a page is not within a namespace and we try to display the
namespace user cap alert, we would receive an error for building the
`pending_members_group_usage_quotas_path`.

We need to use `.persisted?` because by default, we build a new
Namespace record and pass it along to the frontend.