Added intents.json to act as a multiplier to the emojis score
When the emoji are retrieved, it will apply the intents multiplier
to the score if one exists. This will either decrease or increase the
score to force it to float to the top or the bottom of the pile.
When a multiplier doesn't exist the code multiplies by 1.

Updated the awards logic to do the same thing but in jQuery.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/88407