Migrate "Expand" MR sidebar button to ButtonComponent
This migrates the "Expand" MR sidebar button to the ButtonComponent
ViewComponent. This results in a minor visual change as the
angle-double-left icon now inherits the button's size, which is slightly
bigger than the previous s12 icon size.
Though, this does not warrant a changelog entry since the button is
currently behind a disabled-by-default feature flag.