Fix static translations in integrations
We need to wrap these in a proc, otherwise they're evaluated when
the class is loaded and will always be in English.

For the CI integrations this was an oversight when we refactored them
to use the DSL in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/86453.

The validation message in Jira was a pre-existing bug which we noticed
due to the improvements to the `StaticTranslationDefinition` cop in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89945.

This also adds missing specs for all Jira validations.

Changelog: fixed