Create coverage report on root pipeline
This adds full support for coverage report
on pipelines with descendants. Each root pipeline
in a pipeline hierarchy has a single coverage report
stored in a Pipeline Artifact.

To ensure that all jobs that would contribute
to a coverage report is accounted for,
the coverage report is only generated after
the last pipeline in the hierarhy has completed.

In addition, only one coverage report is kept
for each root pipeline. When a coverage report
is recreated for a pipeline, such as when
a failed job is retried or a manual job is triggered,
the new coverage report will be saved in a
new Pipeline Artifact for coverage report
and the existing Pipeline Artifact will be removed.