Exclude expired subscriptions from usage alerts
We don't want to try render seat count alerts for expired subscriptions
as this will result in a large volume of API calls to the CustomersDot
application that are un-needed. Early return in the service that
calculates the seat usage data so that we don't get to the API call
stage. Also exclude trials, as these won't be eligible for alerts
either, so we don't need to make the API call.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/348481
EE: true