Update specs for integrations API to check all modifiable fields
Previously only the event booleans were verified, but not any other
attributes that were sent in the PUT request.

This is a step towards fixing inconsistencies in the Integrations API,
some attributes are currently not supported and skipped in the specs.

This change also exposed a bug in the getters defined by `data_field`,
where we'd incorrectly fallback to `properties` for falsy values. This
should only be done for explicit `nil` values.

Changelog: changed