CloudSQL Instance Worker
This commit introduces the `CreateCloudsqlInstanceWorker`.

Creating Cloud SQL instances is an async operation that
takes a few minutes to conclude.

Thus the Google Cloud API needs to be watched for operation
result and on success an app database and app user must be
created for the new instance.

Worker performs these tasks. Methods added to the Google Cloud
Client to support `get_instance`, `insert_database` and
`insert_user` operations.

`SetupCloudsqlInstanceService` that's invoked from
`CreateCloudsqlInstanceWorker` performs this function.