Change destination_name to destination_slug
To avoid confusion, this updates the API documentation of
bulk imports to make clear that 'destination_name' actually
refers to the destination slug.

We will eventually update the backend to change 'destination_name'
to 'destination_slug', but thought that updating the
documentation now was a good initial step.

Changelog: other