Revert "Use OS tempdir for artifact metadata"
This reverts commit a5b0a030c38acd6da5f66f094a40882f7d2eb892.

The Rails multipart middleware only allows temporary uploads to be
stored in certain paths. This middleware includes Ruby's `Dir.tmpdir`,
but this method may not necessarily return `/tmp`. For example, if the
permissions of `/tmp` do not contain the sticky bit, Ruby may return
the current working directory. In addition, the problem will also show
up if Workhorse and GitLab Rails have `TMPDIR` set differently.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87255 changed
the behavior so that artifact metadata is always stored in
`/tmp`. When Ruby's `Dir.tmpdir` differs from that path, then artifact
uploads will be rejected with a 400 Bad Request.

To avoid this issue, we revert the changes in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87255 and always
take the recommended temp path from Rails.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/363701