Activate users after plan upgrade
With the `free_user_cap`, we cap free groups by setting users to
`awaiting`. However, when a group upgrads to a paid plan, we require
them to purchase seats also for the `awaiting` users.

To gain access again immediately after the upgrade, we activate these
users. However, we also make sure that there are enough seats remaining.
If it's not the case, we no-op. Owners of the group then still have the
ability to activate specific users on the Usage Quota page.