Disable Fips/OpenSSL Rubocop by default
With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/88469, the
Ruby `Digest` modules will automatically use the OpenSSL equivalents
in FIPS mode. We don't need to enforce this as a static lint, nor can
we reasonably expect all upstream dependencies to do the same.

There may also be slight advantange for the Ruby code to run without
the need for OpenSSL.