Add web hook rate limiting columns to database
Note, because we enforce a `Changelog` trailer for database changes,
this title has been written to mention only the database changes as the
feature changes are behind a feature flag.

Real commit title: Log webhook rate limits for paid plans

Currently only customers on the Free Plan have rate limited webhooks.

This change introduces logging of rate limited webhooks of customers
on paid plans.

The limits are weighted according to the number of customer seats,
allowing the limit to scale as the customer grows.

The logging will allow us to assess the limits. A future change will
remove the logging and implement actual rate limiting.

Changelog: added

https://gitlab.com/gitlab-org/gitlab/-/issues/337228