Remove omit_epic_subscribed feature flag
This feature flag was used to gate a change that removed the
subscribed key from the hash embedded in the page when viewing
an epic. This was an expensive operation that is now redundant
as subscribed behavior is handled asynchronously.

Changelog: other
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/88869
EE: true