Fix deploy boards for Kubernetes 1.22
The extensions/v1beta1 was removed on Kubernetes 1.22: https://kubernetes.io/blog/2021/07/14/upcoming-changes-in-kubernetes-1-22/
Whenever we call extensions_client.discover the code raises a 404 not found and we fail to load deployments and ingresses.
Consequently our deploy_boards also fail.

The fix is simply to not use this client anymore, since we don't need to look for resources in this APIs anymore.
We should simply use the new APIs as suggested by the Kubernetes 1.22 documentation.
That is networking.k8s.io/v1 for ingresses and apps/v1 for deployments.

Changelog: fixed