Rescue CustomersDot seat alert connection errors
Currently if the HTTP call to customers dot fails with an exception,
this stops the rendering of the page. As the customers app can go down
for maintenance or other errors, we want to ensure this does not affect
GitLab, as the alert is rendered on a lot of high traffic pages.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/348481
EE: true