Prefer non-expired licenses first
When loading the current license, the first valid and started license is
used. This license can be an expired one while the system also has a
renewal license applied.
This change will look for a license that is valid, started and unexpired
before falling back to looking for a valid and started one if none was
found.

Changelog: fixed
EE: true