Query `vulnerability_reads` by the cartesian product of given filters
SQL queries using IN filters are running quite slow even if we have
index coverage for all the columns involved. To fix this issue, we are
now changing our queries to generate resultset containing the records
for the cartesian product of the given filters to utilize the indices
better.

This will cause read amplification but we are limiting the intermediate
query so the application will not receive all the records.

This feature will be controlled by the `use_unnested_queries` feature
flag for now.

Changelog: added
EE: true