Remove channel fields for chat integrations that don't support them
With some integrations the webhook is already tied to a specific
channel, for others the channels are configurable for each event.

Previously this was weirdly handled by overriding the `event_field`
method in the integrations that don't support channel fields, which
meant the fields were still present in `#fields` but not displayed by
the frontend (controlled by `Integrations::EventEntity`).

To clean this up we're introducing a new method `configurable_channels?`
and only include the channel fields if they're supported.