Fix import/no-unresolved violations
This change makes sure that these import paths are checked by the
import/no-unresolved ESLint rule by making sure that the path resolves
to an actual file when FOSS_ONLY=1.

Since the files are empty, it makes sense not to attempt to fetch them,
which is why they're still guarded by an IS_EE check.