Adjust migration update query and refactor specs
Query used to determine which commits to update was
changed to only update commits which are missing the
project permission fields being added. Refactor the
specs to test the batching and validate the change
works. The new spec will fail without the changes
to the migration.

Changelog: changed
EE: true