Replace jQuery in app/assets/javascripts/projects/star.js file
As part of our effort to migrate away from jquery, we should replace
usage of jquery in the app/assets/javascripts/projects/star.js

Most of the usage in this file is for selecting specific DOM elements or
performing simple operations.

This commit replaces the use of jQuery in
app/assets/javascripts/projects/star.js with vanilla JavaScript.