Add pull-policy support for service images on CI config
In the runner config.toml file, you can already define pull_policy
for runners. This config determines how to fetch the needed images.
In this change, we are adding this config to CI config YAML.
The config can be a string or array of strings.
The config will be filtered by the allowed_pull_policies config
in the Runner side.

We already did this for job images.

This is behind a feature flag ci_docker_image_pull_policy