Add sbom_sources table and model
This adds a new table named `sbom_sources` along with its associated
model. This is part two of a three-phase implementation for adding new
models to support the tracking of Software Bill of Materials (SBoM)
components in GitLab. `sbom_sources` will be used in order to track the
location in which an SBoM component originated from. For example,
library components are typically detected from a dependency file, such
as `package-lock.json`.

Changelog: added