Add quick action support for group members
Prior to this change, we had removed support for assignment to groups
in quick actions. This was a largely accidental feature, but widely
used by power users. This change re-introduces support for assigment
(and other actions) on group members.

Rather than a simple revert, this improves the former accidental
feature, with:

- typo detection: groups that don't resolve are an error
- feature level support: group assigment is only available where
  multiple assigment is supported.

Changelog: changed