Fix future iterations not being scheduled correctly
This commit updates ScheduleCreateIterationsWorker to
use the correct scope to pick up the cadences that need
to have future iterations scheduled. After the update,
the worker will pick up the cadences whose `last_run_date`
is earlier than or is on the current date.

To ensure every automated cadence has the required number
of future iterations, the commit adds a post-deployment migration
to reset `last_run_date` to the current date. This will allow
ScheduleCreateIterationsWorker to iterate through every automated
cadence.

Changelog: fixed