Refactor CI rules entry to use composed value
When we return a value from an entry, we always need to return the
composed version. In Rules, we return "config", which means the hash
version of what the YAML file has. I changed it to "super" because
it needs to have composed value of each sub entry.

With this, we also needed to change the
validate_against_warnings method. Since we haven't composed the nested
entries yet, we cannot use *_value attributes there.

These changes are behind a feature flag
ci_value_change_for_processable_and_rules_entry