Change the revoke button to type `tertiary`
We decided this change was better than these two other proposals:

* [Draft: Change the revoke button type to `secondary`](
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89504
)

* [Draft: Change the revoke button to 3 different types](
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/89507
)

In the AccessTokenTableApp component, users may not understand why
some access tokens have a primary revoke button (background red),
while other access tokens have a secondary (transparent background).

Originally, I believe the primary|secondary buttons were meant to
match the color of the text of the Expires column (3 colors: red for
expired token, orange for soon-to-be-expired, black for tokens without
expiration). However, the match for the button was not perfect:
primary button for tokens with expiration, secondary button for
tokens without expiration.

This MR changes the revoke buttons to be of type tertiary.

Changelog: changed