Start emitting gitlab_sli metrics separated with '_'
This is the first step in switching to metrics separated with an `_`
as recommended in the Prometheus best practices:

> Note: The colons are reserved for user defined recording rules. They
> should not be used by exporters or direct instrumentation.

https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels

With this, we can optionally start emitting the new metrics next to
the old ones. This allows us to keep recording the old ones for now,
since they are used for alerting and error budgets for stage
groups. But the new ones will be present so we can make a switch.