Fix broken Prometheus alert processing test
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/90289 exposed a
test set up problem with
`spec/services/projects/prometheus/alerts/notify_service_spec.rb`.

The test was failing because it expected 2 alerts to be processed, but
only 1 was returned. This was happening because the resolved alert
could not be found in the database, and since the incoming alert
payload was already resolved, a new alert was not persisted again.

In the test, we have two Prometheus alerts:

1. `alert_firing`
2. `alert_resolved`

In a clean database state:

1. `alert_resolved.id` = 2
2. `alert_firing.prometheus_metric_id` = 1
3. `alert_resolved.prometheus_metric_id` = 2

If `alert_resolved.id` happens to match
`alert_resolved.prometheus_metric_id`, the test passes.

But if you run a spec that increments the counters for the
`prometheus_alerts` or `alert_management_alerts` tables (such as
`spec/models/factories_spec.rb`), then the IDs will diverge. This
leads to the test failure since the fingerprint lookup will fail to
match since it is calculated with the wrong ID.

The fingerprint calculation should use `prometheus_metric_id`, not
`id`.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/366261