Use UTC for trial end date in popover
Because we create a `Date` Object and pass it to the frontend, the end
date of the trial is shown in the browser's timezone.
However, we store the end date in UTC and also show the end date in UTC
on the billing page. To streamline the dates, this MR changes the
formatted date in the trial popover to be in UTC as well.

Changelog: fixed
EE: True