Add post-migration test table
As part of https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/585, we
need to test the execution of post-migrations in the post-deploy
pipeline. Given post-migrations are not common, we decided to test with
a fake post-migration, this commit introduces one that creates a fake
table, the table is going to be removed after the testing is done.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2352

Changelog: other