gitaly_client: Fix error handling for structured AlreadyApplied error
With 21d14354183b (Add structured error parsing for user_cherry_pick,
2022-05-04) we have introduced error handling for structuder errors
returned by UserCherryPick. This new mode of error handling is slated to
replace the old error reporting in Gitaly, which used to embed errors in
a successful UserCherryPikc response.

This new error handling is missing the case where the commits we're
about to cherry-pick have already been applied to the target branch and
are thus empty, which is breaking some of our tests when upgrading to a
Gitaly version that already has the structured error reporting.

Fix this by handling `ChangesAlreadyApplied` errors equivalently to the
previous `CreateTreeError` with its error code set to `EMPTY`.

Changelog: fixed